requirements.txt
zhconv
nuitka
natsort

pdf2image
pillow

alive-progress

tomli
tomli_w
